removed unessecary files and setup websocket server independent of udp sockets
