+++ .gitignore
@@ -2,3 +2,5 @@
 install/
 build/
 log/
+
+node_modules/
